Fixes #51 (in github)
Access key of 'k' added to the toolbarbutton.  This triggers whatever is in the 'oncommand' attribute, so function set up to handle it.

Both the button's 'oncommand' handler and a mouse click on the button show the associated popup.  So instead of having .click() handlers, we use the
button's 'onpopupshowing' and 'onpopupshown' attributes to call .prePopup() and .postPopup() for grabbing data on already bookmarked pages, setting initial focus,
etc.  This avoids the bug where the handler for .click() on the button was called for any click inside the popup panel.

Also set up <command> entities in bookie.xul to handle the Save and Delete buttons.  This way, a user can tab to either button and hit the spacebar or
whatever to activate it instead of having to click on it with the mouse.